# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/landing/{code}/banners": {
            "get": {
                "description": "Get Landing Banners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Landing"
                ],
                "summary": "Get Landing Banners",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/landing.GetLandingBannersResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "landing.Banner": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "start_date": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "landing.GetLandingBannersResponse": {
            "type": "object",
            "properties": {
                "banners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/landing.Banner"
                    }
                }
            }
        },
...
    }
}